% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{RakeRvis}
\alias{RakeRvis}
\title{Shiny app to launch an interactive session}
\usage{
RakeRvis(
  data = NULL,
  Providers = c("OpenStreetMap", "GoogleEarth", "Esri.OceanBasemap", "Esri.WorldImagery"),
  browser = TRUE,
  TagID = NULL,
  col = NULL,
  col_anim = NULL,
  plot_w = "100\%",
  plot_h = 800,
  shapes = NULL,
  pcol = "black",
  scol = "black",
  points = NULL,
  PathOut = NULL,
  legend = TRUE,
  lines4fixes = TRUE,
  fixes = TRUE
)
}
\value{
A nice Shiny browser of goodness
}
\description{
Interactive Shiny App for MoveRakeR visualisation
This app is still in beta testing development.
}
\details{
More to add here, placeholder text
}
